requirements.txt
matplotlib==3.8.4
numpy==1.26.4
Pillow==10.3.0
